Refactor annotations package structure and add missing request mapping in Settings API interface
